Rename MoreBottomSheet -> BottomSheet
