+++ client/src/components/BottomSheet/index.tsx
@@ -1,0 +1,2 @@
+export { default } from "./BottomSheet";
+export * from "./BottomSheet";
