--- client/src/components/MoreBottomSheet/index.tsx
@@ -1,2 +1,0 @@
-export { default } from "./MoreBottomSheet";
-export * from "./MoreBottomSheet";
